build: introduce microbundle
