+++ .github/workflows/node.js.yml
@@ -25,4 +25,5 @@ jobs:
         node-version: ${{ matrix.node-version }}
     - run: yarn install --frozen-lockfile
     - run: yarn lint
+    - run: yarn build
     - run: yarn test
