+++ src/xmlhttprequesteventtarget.js
@@ -1,7 +1,7 @@
 /**
  * The MIT License (MIT)
  *
- * Copyright (c) 2013-2017 Yamagishi Kazutoshi
+ * Copyright (c) 2012-2020 Yamagishi Kazutoshi
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -22,22 +22,7 @@
  * THE SOFTWARE.
  */
 
-'use strict';
+import EventTarget from './eventtarget';
 
-class DOMException extends Error {
+export default class XMLHttpRequestEventTarget extends EventTarget {
 }
-
-(function() {
-  const domExceptionConstrants = {
-    INVALID_STATE_ERROR: {
-      configurable: false,
-      enumerable: true,
-      value: 11,
-      writable: false
-    }
-  };
-  Object.defineProperties(DOMException, domExceptionConstrants);
-  Object.defineProperties(DOMException.prototype, domExceptionConstrants);
-})();
-
-module.exports = DOMException;
